[RENAME/#111] explore->product 패키지명 변경
