+++ app/src/main/java/com/napzak/market/data/product/dto/request/ProductBuyItemsRequest.kt
@@ -1,20 +1,16 @@
-package com.napzak.market.data.explore.dto.request
+package com.napzak.market.data.product.dto.request
 
 import kotlinx.serialization.SerialName
 import kotlinx.serialization.Serializable
 
 @Serializable
-data class SearchedProductSellItemsRequest(
-    @SerialName("searchWord")
-    val searchWord: String,
+data class ProductBuyItemsRequest(
     @SerialName("sortOption")
     val sortOption: String,
     @SerialName("genreId")
     val genreIds: List<Long>?,
     @SerialName("isOnSale")
     val isOnSale: Boolean,
-    @SerialName("isUnopened")
-    val isUnopened: Boolean,
     @SerialName("nextCursor")
     val nextCursor: String = "",
 )
