+++ app/src/main/java/com/napzak/market/data/product/dto/request/SearchedProductBuyItemsRequest.kt
@@ -1,18 +1,18 @@
-package com.napzak.market.data.explore.dto.request
+package com.napzak.market.data.product.dto.request
 
 import kotlinx.serialization.SerialName
 import kotlinx.serialization.Serializable
 
 @Serializable
-data class ProductSellItemsRequest(
+data class SearchedProductBuyItemsRequest(
+    @SerialName("searchWord")
+    val searchWord: String,
     @SerialName("sortOption")
     val sortOption: String,
     @SerialName("genreId")
     val genreIds: List<Long>?,
     @SerialName("isOnSale")
     val isOnSale: Boolean,
-    @SerialName("isUnopened")
-    val isUnopened: Boolean,
     @SerialName("nextCursor")
     val nextCursor: String = "",
 )
