+++ app/src/main/java/com/napzak/market/data/product/mapper/ExploreProductSellItemsMapper.kt
@@ -1,7 +1,7 @@
-package com.napzak.market.data.explore.mapper
+package com.napzak.market.data.product.mapper
 
-import com.napzak.market.data.explore.dto.request.ProductSellItemsRequest
-import com.napzak.market.data.explore.dto.response.ProductSellItemsResponse
+import com.napzak.market.data.product.dto.request.ProductSellItemsRequest
+import com.napzak.market.data.product.dto.response.ProductSellItemsResponse
 import com.napzak.market.domain.explore.model.ProductItem
 import com.napzak.market.domain.explore.model.ProductListFilter
