+++ app/src/main/java/com/napzak/market/data/product/mapper/ExploreSearchedBuyItemsMapper.kt
@@ -1,6 +1,6 @@
-package com.napzak.market.data.explore.mapper
+package com.napzak.market.data.product.mapper
 
-import com.napzak.market.data.explore.dto.request.SearchedProductBuyItemsRequest
+import com.napzak.market.data.product.dto.request.SearchedProductBuyItemsRequest
 import com.napzak.market.domain.explore.model.ProductListFilterWithSearchTerm
 
 fun ProductListFilterWithSearchTerm.toSearchedProductBuyRequest(): SearchedProductBuyItemsRequest =
